Created new teleporting feature
