+++ src/TodoList.js
@@ -3,11 +3,19 @@ import TodoListItem from "./TodoListItem"
 
 //function to assemble and dissemble list.
 
-const TodoList = ({ todoList, onRemoveTodo }) => {
+const TodoList = ({ todoList, onRemoveTodo, setCurrentTodo, currentTodo }) => {
+
+
   return (
     <ul>
       {todoList.map((todo) => (
-      <TodoListItem key={todo.id} todo={todo} onRemoveTodo={onRemoveTodo} />
+        <TodoListItem
+          key={todo.id}
+          todo={todo}
+          onRemoveTodo={onRemoveTodo}
+          setCurrentTodo={setCurrentTodo}
+          currentTodo={currentTodo}
+        />
       ))}
     </ul>
   );
