+++ src/TodoListItem.js
@@ -1,11 +1,17 @@
 import React, { useState } from "react";
-import TodoList from "./TodoList";
-import AddTodoForm from "./Forms.js/AddTodoForm";
-import InputWithLabel from "./Forms.js/AddTodoLabel";
+import EditForm from "./Forms.js/EditForm";
 
 //Function creates list item with remove button. Add checkbox, edit, favorite here.
 
-const TodoListItem = ({ todo, onRemoveTodo, onEditClick}) => {
+const TodoListItem = ({ todo, onRemoveTodo, setCurrentTodo, currentTodo, setTodoList }) => {
+  
+  const [isEditing, setIsEditing] = useState(false);
+
+  const handleEditClick = (todoTitle) => {
+    setIsEditing(true);
+    setCurrentTodo({ ...todoTitle });
+  };
+  
   console.log(todo.id);
   return (
       
@@ -13,7 +19,18 @@ const TodoListItem = ({ todo, onRemoveTodo, onEditClick}) => {
       
       {todo.title}
 
-      <button type="button" onClick={() => onEditClick(todo.title)}>Edit</button>
+      {!isEditing ?
+        <button type="button" onClick={() => handleEditClick(todo.title)}>Edit</button>
+        :
+        <EditForm
+          setTodoList={setTodoList}
+          setCurrentTodo={setCurrentTodo}
+          currentTodo={currentTodo}
+          setIsEditing={setIsEditing} 
+          />
+      }
+
+      
       <button type="button" onClick={() => onRemoveTodo(todo.id)}>Remove Button</button>
       
     </li>
